requirements.txt
streamlit
pandas
numpy
scipy
plotly
python-dotenv
ccxt